requirements.txt
python-dotenv~=0.19.2
git+https://github.com/Pycord-Development/pycord